% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{stack_view}
\alias{stack_view}
\title{See what's on the stack}
\usage{
stack_view(x)
}
\arguments{
\item{x}{optional string. If supplied it should match the name of an object in the package enviroment.
The value of the corresponding variable will be returned. If missing, a list of all objects in the package enviroment.}
}
\description{
This is primarily to help with debugging.
}
\note{
Note that \code{\link{graft}} does not delete objects from the environment.
}
