% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabe.R
\name{clear_stack}
\alias{clear_stack}
\title{Reset the stack so that objects left from previous pipelines will not cause issues}
\usage{
clear_stack()
}
\description{
Reset the stack so that objects left from previous pipelines will not cause issues
}

