% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabe.R
\name{clear_stack}
\alias{clear_stack}
\title{Remove all objects from the stack by deleting them from memory.}
\usage{
clear_stack()
}
\description{
Remove all objects from the stack by deleting them from memory.
}

